requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
matplotlib>=3.8.0
yfinance>=0.2.54
PyYAML>=6.0.0
pytest>=8.0.0
requests>=2.31.0
streamlit>=1.42.0
